+++ zerver/worker/base.py
@@ -16,7 +16,7 @@ from returns.curry import partial
 from typing_extensions import override
 
 from zerver.lib.context_managers import lockfile
-from zerver.lib.db import reset_queries
+from zerver.lib.db_connections import reset_queries
 from zerver.lib.per_request_cache import flush_per_request_caches
 from zerver.lib.pysa import mark_sanitized
 from zerver.lib.queue import SimpleQueueClient
